requirements.txt
flask
mysqldb
google-cloud
google-api-python-client
google-auth-httplib2
gcloud sdk
google-cloud-spanner
